+++ frontend/src/FullPageRoutes/Profile.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
 import { request } from "../utils/fetch"
 import { getUpdates, useProfile } from "../utils/profile"
-import ProfileSidebar from "../components/ProfileSidebar"
-import ProfileContent from "../components/ProfileContent"
-import type { ProfileContentType } from "../components/ProfileContent"
+import ProfileSidebar from "../components/profile/ProfileSidebar"
+import ProfileContent from "../components/profile/ProfileContent"
+import type { ProfileContentType } from "../components/profile/ProfileContent"
 import { makeStyles } from "@material-ui/core/styles"
 import { Grid } from "@material-ui/core"
 import { Nullable } from "../utils/context"
