+++ frontend/src/components/profile/ProfileSidebar.tsx
